explation + refactor
