+++ event-based-asynchronous/pom.xml
@@ -32,7 +32,7 @@
     <artifactId>java-design-patterns</artifactId>
     <version>1.26.0-SNAPSHOT</version>
   </parent>
-  <artifactId>event-asynchronous</artifactId>
+  <artifactId>event-based-asynchronous</artifactId>
   <dependencies>
     <dependency>
       <groupId>org.junit.jupiter</groupId>
