+++ event-based-asynchronous/src/main/java/com/iluwatar/event/asynchronous/Event.java
