+++ event-based-asynchronous/src/main/java/com/iluwatar/event/asynchronous/InvalidOperationException.java
@@ -24,11 +24,14 @@
  */
 package com.iluwatar.event.asynchronous;
 
+import java.io.Serial;
+
 /**
  * Type of Exception raised when the Operation being invoked is Invalid.
  */
 public class InvalidOperationException extends Exception {
 
+  @Serial
   private static final long serialVersionUID = -6191545255213410803L;
 
   public InvalidOperationException(String message) {
