+++ event-based-asynchronous/src/main/java/com/iluwatar/event/asynchronous/LongRunningEventException.java
@@ -24,11 +24,14 @@
  */
 package com.iluwatar.event.asynchronous;
 
+import java.io.Serial;
+
 /**
  * Type of Exception raised when the Operation being invoked is Long Running.
  */
 public class LongRunningEventException extends Exception {
 
+  @Serial
   private static final long serialVersionUID = -483423544320148809L;
 
   public LongRunningEventException(String message) {
