+++ pom.xml
@@ -140,7 +140,7 @@
     <module>promise</module>
     <module>page-controller</module>
     <module>page-object</module>
-    <module>event-asynchronous</module>
+    <module>event-based-asynchronous</module>
     <module>event-queue</module>
     <module>queue-load-leveling</module>
     <module>object-mother</module>
